合并最新代码

+++ shop-api/src/main/java/cn/lili/consumer/listener/GoodsMessageListener.java
@@ -8,7 +8,7 @@ import cn.hutool.json.JSONObject;
 import cn.hutool.json.JSONUtil;
 import cn.lili.consumer.event.GoodsCommentCompleteEvent;
 import cn.lili.modules.distribution.entity.dos.DistributionGoods;
-import cn.lili.modules.distribution.entity.dos.DistributionSelectedGoods;
+import cn.lili.modules.distribution.entity.dto.DistributionGoodsSearchParams;
 import cn.lili.modules.distribution.service.DistributionGoodsService;
 import cn.lili.modules.distribution.service.DistributionSelectedGoodsService;
 import cn.lili.modules.goods.entity.dos.*;
@@ -27,12 +27,11 @@ import cn.lili.modules.promotion.entity.dto.BasePromotions;
 import cn.lili.modules.promotion.entity.enums.PromotionsScopeTypeEnum;
 import cn.lili.modules.promotion.entity.vos.PromotionGoodsSearchParams;
 import cn.lili.modules.promotion.service.PromotionGoodsService;
+import cn.lili.modules.promotion.service.PromotionService;
 import cn.lili.modules.search.entity.dos.EsGoodsIndex;
 import cn.lili.modules.search.service.EsGoodsIndexService;
 import cn.lili.modules.store.service.StoreService;
 import cn.lili.rocketmq.tags.GoodsTagsEnum;
-import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
-import com.baomidou.mybatisplus.core.conditions.update.LambdaUpdateWrapper;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.rocketmq.common.message.MessageExt;
 import org.apache.rocketmq.spring.annotation.RocketMQMessageListener;
@@ -116,6 +115,9 @@ public class GoodsMessageListener implements RocketMQListener<MessageExt> {
     private StoreGoodsLabelService storeGoodsLabelService;
 
     @Autowired
+    private PromotionService promotionService;
+
+    @Autowired
     private PromotionGoodsService promotionGoodsService;
 
     @Override
@@ -184,12 +186,28 @@ public class GoodsMessageListener implements RocketMQListener<MessageExt> {
                 break;
             //审核商品
             case GOODS_AUDIT:
-                updateGoodsNum(messageExt);
+                Goods goods = JSONUtil.toBean(new String(messageExt.getBody()), Goods.class);
+                updateGoodsNum(goods);
                 break;
             //删除商品
             case GOODS_DELETE:
-                deleteGoods(messageExt);
-                updateGoodsNum(messageExt);
+                try {
+                    String goodsIdsJsonStr = new String(messageExt.getBody());
+                    for (String goodsId : JSONUtil.toList(goodsIdsJsonStr, String.class)) {
+                        Goods goodsById = this.goodsService.getById(goodsId);
+                        if (goodsById != null) {
+                            this.deleteGoods(goodsById);
+                            this.updateGoodsNum(goodsById);
+                            List<String> skuIdsByGoodsId = this.goodsSkuService.getSkuIdsByGoodsId(goodsId);
+                            if (skuIdsByGoodsId != null && !skuIdsByGoodsId.isEmpty()) {
+                                this.goodsIndexService.deleteIndexByIds(skuIdsByGoodsId);
+                            }
+                        }
+                    }
+
+                } catch (Exception e) {
+                    log.error("删除商品索引事件执行异常，商品信息 {}", new String(messageExt.getBody()));
+                }
                 break;
             //规格删除
             case SKU_DELETE:
@@ -258,7 +276,10 @@ public class GoodsMessageListener implements RocketMQListener<MessageExt> {
         List<EsGoodsIndex> goodsIndices = new ArrayList<>();
         for (Goods goods : goodsList) {
             //如果商品通过审核&&并且已上架
-            List<GoodsSku> goodsSkuList = this.goodsSkuService.list(new LambdaQueryWrapper<GoodsSku>().eq(GoodsSku::getGoodsId, goods.getId()).gt(GoodsSku::getQuantity, 0));
+            GoodsSearchParams searchParams = new GoodsSearchParams();
+            searchParams.setGoodsId(goods.getId());
+            searchParams.setGeQuantity(0);
+            List<GoodsSku> goodsSkuList = this.goodsSkuService.getGoodsSkuByList(searchParams);
             if (goods.getAuthFlag().equals(GoodsAuthEnum.PASS.name())
                     && goods.getMarketEnable().equals(GoodsStatusEnum.UPPER.name())
                     && Boolean.FALSE.equals(goods.getDeleteFlag())) {
@@ -287,7 +308,9 @@ public class GoodsMessageListener implements RocketMQListener<MessageExt> {
      */
     private void updateGoodsIndex(Goods goods) {
         //如果商品通过审核&&并且已上架
-        List<GoodsSku> goodsSkuList = this.goodsSkuService.list(new LambdaQueryWrapper<GoodsSku>().eq(GoodsSku::getGoodsId, goods.getId()));
+        GoodsSearchParams searchParams = new GoodsSearchParams();
+        searchParams.setGoodsId(goods.getId());
+        List<GoodsSku> goodsSkuList = this.goodsSkuService.getGoodsSkuByList(searchParams);
         if (goods.getAuthFlag().equals(GoodsAuthEnum.PASS.name())
                 && goods.getMarketEnable().equals(GoodsStatusEnum.UPPER.name())
                 && Boolean.FALSE.equals(goods.getDeleteFlag())) {
@@ -359,6 +382,11 @@ public class GoodsMessageListener implements RocketMQListener<MessageExt> {
                 goodsIndex.setStoreCategoryNamePath(ArrayUtil.join(storeGoodsLabels.stream().map(StoreGoodsLabel::getLabelName).toArray(), ","));
             }
         }
+
+        if (goodsIndex.getPromotionMap() == null || goodsIndex.getPromotionMap().isEmpty()) {
+            Map<String, Object> goodsCurrentPromotionMap = promotionService.getGoodsPromotionMap(goodsIndex);
+            goodsIndex.setPromotionMap(goodsCurrentPromotionMap);
+        }
     }
 
 
@@ -368,37 +396,37 @@ public class GoodsMessageListener implements RocketMQListener<MessageExt> {
      * 2.删除分销员-分销商品绑定关系
      * 3.删除分销商品
      *
-     * @param messageExt 消息
+     * @param goods 消息
      */
-    private void deleteGoods(MessageExt messageExt) {
-        Goods goods = JSONUtil.toBean(new String(messageExt.getBody()), Goods.class);
+    private void deleteGoods(Goods goods) {
 
+        DistributionGoodsSearchParams searchParams = new DistributionGoodsSearchParams();
+        searchParams.setGoodsId(goods.getId());
         //删除获取分销商品
-        DistributionGoods distributionGoods = distributionGoodsService.getOne(new LambdaQueryWrapper<DistributionGoods>()
-                .eq(DistributionGoods::getGoodsId, goods.getId()));
+        DistributionGoods distributionGoods = distributionGoodsService.getDistributionGoods(searchParams);
 
-        //删除分销商品绑定关系
-        distributionSelectedGoodsService.remove(new LambdaQueryWrapper<DistributionSelectedGoods>()
-                .eq(DistributionSelectedGoods::getDistributionGoodsId, distributionGoods.getId()));
+        if (distributionGoods != null) {
 
-        //删除分销商品
-        distributionGoodsService.removeById(distributionGoods.getId());
+            //删除分销商品绑定关系
+            distributionSelectedGoodsService.deleteByDistributionGoodsId(distributionGoods.getId());
+
+            //删除分销商品
+            distributionGoodsService.removeById(distributionGoods.getId());
+        }
     }
 
     /**
      * 修改商品数量
      *
-     * @param messageExt 信息体
+     * @param goods 信息体
      */
-    private void updateGoodsNum(MessageExt messageExt) {
-
+    private void updateGoodsNum(Goods goods) {
         try {
-            Goods goods = JSONUtil.toBean(new String(messageExt.getBody()), Goods.class);
             //更新店铺商品数量
             assert goods != null;
             storeService.updateStoreGoodsNum(goods.getStoreId());
         } catch (Exception e) {
-            log.error("商品MQ信息错误：{}", messageExt.toString());
+            log.error("修改商品数量错误");
         }
     }
 
@@ -421,9 +449,7 @@ public class GoodsMessageListener implements RocketMQListener<MessageExt> {
                     goods.setBuyCount(0);
                 }
                 int buyCount = goods.getBuyCount() + goodsCompleteMessage.getBuyNum();
-                goodsService.update(new LambdaUpdateWrapper<Goods>()
-                        .eq(Goods::getId, goodsCompleteMessage.getGoodsId())
-                        .set(Goods::getBuyCount, buyCount));
+                this.goodsService.updateGoodsBuyCount(goodsCompleteMessage.getGoodsId(), buyCount);
             } else {
                 log.error("商品Id为[" + goodsCompleteMessage.getGoodsId() + "的商品不存在，更新商品失败！");
             }
